requirements.txt
pip==21.1.3
bump2version==1.0.1
wheel==0.36.2
watchdog==2.1.3
flake8==3.9.2
tox==3.23.1
coverage==5.5
Sphinx==4.0.2
twine==3.4.1
Click==8.0.1
dependency_injector==4.34.0

setuptools~=52.0.0
pyyaml
python-dotenv
